wordpress.py: Increase per_page value to fix pulling plugins
WordPress.org's API has this weird behavior with plugins where when
pulling the default of 25 items per page, the first three pages are
structured differently than the rest of the pages. It starts with the
plugin data being in an array and then switches to an indexed
dictionary after page 3. By now pulling 100 items per_page, this is no
longer the case and data can be pulled reliably.

The threading model is still pretty juvenile, where each page is
retrieved and the items on the page are used to perform a more
detailed lookup in parallel and joined before proceeding to the next
page. This results in inconsistent resource usage, though due to
coincidence this doesn't violate the API requests per second
limitations.