envato.py: Collect more accurate data
Before I was assuming that I was getting all the data from the category
I had selected though this was false. Because authors may be marketing
their themes to multiple categories this created duplicate entries. Now
I have set the default category to 'wordpress/corporate' which is the
largest category for wordpress on themeforest, and set the items per
page to be 100. Because there is a limit of 60 pages and up to 100 items
per page from the API, you can at most collect 6000 items from the api
through pagination. Luckily most categories have fewer than half of this
limit, so my changes allow creating exports of individual categories
that the user may review and/or combine.